create field ok, arrumando migration da tabela field

+++ backend/src/database/migrations/1609225348858-createFieldTable.ts
@@ -40,12 +40,12 @@ export default class createFieldTable1609225348858
           },
           {
             name: "latitude",
-            type: "int",
+            type: "float",
             isNullable: false,
           },
           {
             name: "longitude",
-            type: "int",
+            type: "float",
             isNullable: false,
           },
           {
